+++ src/controllers/StencilsController.php
@@ -119,7 +119,6 @@ class StencilsController extends Controller
         return $this->renderTemplate('formie/stencils/_edit', [
             'stencil' => $stencil,
             'canEdit' => $stencil->canEdit(),
-            'builderSiteCrumb' => Formie::$plugin->getFormSiteOverrides()->getBuilderSiteCrumbConfig(),
         ]);
     }